Changes to MIME handling and @meta_data propery
getDocumentAsync now has an option parameter, which can specify which items to return.
getDocumentAsync also returns a @meta_data propery with the data type of the item. (number,string,mime, array). Also for mime items, it also returns the mime header.

the getMimeItem() functions now  returns a object with the mime value and mime header.